almost implemented the game
player cannot deploy ships
TODO: add a database
