+++ assignment1_part2/src/Controller/PlayerDisplayListener.java
@@ -5,7 +5,6 @@
  */
 package Controller;
 
-import View.PlayerDisplay;
 import java.awt.event.MouseEvent;
 
 /**
