+++ assignment1_part2/src/View/ShipListPanel.java
@@ -5,6 +5,7 @@
  */
 package View;
 
+import Controller.PlayerDisplay;
 import Model.Game;
 import Model.Player;
 import Model.Ship;
@@ -35,16 +36,15 @@ public class ShipListPanel extends JPanel {
 
     private void constructGUI() {
         JPanel shipListPanel = new JPanel();
-        {
-            shipListPanel.setLayout(new FlowLayout());
 
-            Ship[] ships = player.getShips();
-            allShipLabels = new JLabel[ships.length];
-            for (int i = 0; i < allShipLabels.length; ++i) {
-                JLabel shipLabel = allShipLabels[i] = new JLabel(ships[i].getName());
-                shipLabel.setBorder(new EmptyBorder(UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH));
-                shipListPanel.add(shipLabel);
-            }
+        shipListPanel.setLayout(new FlowLayout());
+
+        Ship[] ships = player.getShips();
+        allShipLabels = new JLabel[ships.length];
+        for (int i = 0; i < allShipLabels.length; ++i) {
+            JLabel shipLabel = allShipLabels[i] = new JLabel(ships[i].getName());
+            shipLabel.setBorder(new EmptyBorder(UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH));
+            shipListPanel.add(shipLabel);
         }
 
         JPanel gridPanel = new JPanel();
@@ -64,10 +64,10 @@ public class ShipListPanel extends JPanel {
 
         Status = new JLabel("\n");
 
-        Status.setBorder(new EmptyBorder(UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH));
+        Status.setBorder(new EmptyBorder(UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH));
         Status.setHorizontalAlignment(SwingConstants.CENTER);
 
-        contentPanel.setBorder(new EmptyBorder(UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH, UI.BORDER_WIDTH));
+        contentPanel.setBorder(new EmptyBorder(UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH, UIBattleApp.BORDER_WIDTH));
         contentPanel.setLayout(new BorderLayout());
         contentPanel.add(shipListPanel, BorderLayout.NORTH);
         contentPanel.add(gridPanel, BorderLayout.CENTER);
